feat: add test helpers to common
